Вынес типы Place и PlaceCard в отдельный файл
