+++ src/types/place.ts
@@ -1,0 +1,25 @@
+export interface Place {
+	id: string
+	title: string
+	type: string
+	price: number
+	city: {
+		name: string
+		location: {
+			latitude: number
+			longitude: number
+			zoom: number
+		}
+	}
+	location: {
+		latitude: number
+		longitude: number
+		zoom: number
+	}
+	isFavorite: boolean
+	isPremium: boolean
+	rating: number
+	previewImage: string
+}
+
+export type PlaceCard = Omit<Place, 'city' | 'location'>
